:bug: Fixes Vitest Location Bugging
